Clean up macros code so it's easier to understand
